finalizando decimo quinto video

+++ src/services/products.js
@@ -4,6 +4,7 @@ import { handleGetProductLocalStorage, setInLocalStorage } from "../persistence/
 import { closeModel } from "../views/modal";
 import { handleGetProductsToStore } from "../views/store";
 import { productoActivo } from "../../main";
+import Swal from "sweetalert2";
 
 
 export const handleSaveOrModifyElements = () => {
@@ -32,6 +33,11 @@ export const handleSaveOrModifyElements = () => {
         };
     }
 
+    Swal.fire({
+        title: "Listo!",
+        text: "Producto guardado correctamente!",
+        icon: "success"
+    });
 
 
     setInLocalStorage(object);
@@ -40,10 +46,31 @@ export const handleSaveOrModifyElements = () => {
 };
 
 export const handleDeleteProduct = () => {
-    const products = handleGetProductLocalStorage();
-    const result = products.filter((el) => el.id !== productoActivo.id);
-    localStorage.setItem("products", JSON.stringify(result));
-    const newproducts = handleGetProductLocalStorage();
-    handleGetProductsToStore(newproducts);
-    closeModel();
+
+    Swal.fire({
+        title: "Seguro que lo deseas borrar ?",
+        text: "No es una acción reversible",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Si, eliminar"
+    }).then((result) => {
+        if (result.isConfirmed) {
+            const products = handleGetProductLocalStorage();
+            const result = products.filter((el) => el.id !== productoActivo.id);
+            localStorage.setItem("products", JSON.stringify(result));
+            const newproducts = handleGetProductLocalStorage();
+            handleGetProductsToStore(newproducts);
+            closeModel();
+            Swal.fire({
+                title: "Eliminado!",
+                text: "El producto fue eliminado correctamente",
+                icon: "success"
+            });
+        } else {
+            closeModel();
+        }
+    });
+
 }
